Deprecate usage of XML Spring schemas
Move MultipleMappings test

Related to #141.
